Keep the file position in an off_t and allow seeking to the end

+++ include/fs/reg_file.h
@@ -7,7 +7,7 @@
 #include <fs/fs.h>
 
 struct reg_file {
-    int pos;
+    off_t pos;
     struct file file;
 };
